requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.6
selenium==4.18.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
psutil==5.9.8
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
eventlet==0.33.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
requests==2.31.0
cryptography==41.0.5 